feat(endpoint): Add QueueReporter and restructure files in report/
